requirements.txt
fpdf==1.7.2
PyMuPDF==1.22.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.4
scikit-learn==1.4.2
flask